+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to Paradigm will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.44.0] — 2026-03-15
+## [3.45.0] — 2026-03-15
 
 ### Added